pyproject.toml
[project]
name = "symfonitor"
version = "0.1.0"
description = ""
authors = [
    {name = "shoxxdj",email = "shoxx@shoxxdj.fr"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "rich (>=14.2.0,<15.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)"
]

[tool.poetry.scripts]
symfonitor="symfonitor.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
